Add 84 image files for products and categories
Add a total of 84 image files, organized into product and category classifications. Each file is uniquely identified, enhancing the visual assets available for the application. The goal of adding those images because of a developer or client doesn't has a cloudfare token or any image cloud provider it waill fallback to the local storage.